Consolidate watershed overview sections.
