+++ client/src/components/side-panels/sections/RhessysOutputsSection.tsx
@@ -1,22 +1,27 @@
+import { tss } from "../../../utils/tss";
 import { useCallback, useMemo } from "react";
 import Typography from "@mui/material/Typography";
 import CircularProgress from "@mui/material/CircularProgress";
-import type { SelectChangeEvent } from "@mui/material/Select";
+import FormControl from "@mui/material/FormControl";
+import InputLabel from "@mui/material/InputLabel";
+import Select, { type SelectChangeEvent } from "@mui/material/Select";
+import MenuItem from "@mui/material/MenuItem";
+import ToggleButton from "@mui/material/ToggleButton";
+import ToggleButtonGroup from "@mui/material/ToggleButtonGroup";
 import { useWatershed } from "../../../contexts/WatershedContext";
+import { useRunId } from "../../../hooks/useRunId";
+import { useRhessysOutputs } from "../../../hooks/useRhessysOutputs";
+
 import {
   getLayerParams,
   type RhessysOutputParams,
 } from "../../../layers/types";
-import { tss } from "../../../utils/tss";
 
-import type {
-  RhessysOutputScenario,
-  RhessysOutputVariable,
-} from "../../../api/types";
-
-import { GATE_CREEK_VARIABLES } from "../../../api/rhessys/constants";
-import RasterModeSection from "./RasterModeSection";
-import ChoroplethModeSection from "./ChoroplethModeSection";
+import {
+  GATE_CREEK_VARIABLES,
+  GATE_CREEK_SCENARIOS,
+  GATE_CREEK_YEAR_RANGE,
+} from "../../../api/rhessys/constants";
 
 const useStyles = tss.create(({ theme }) => ({
   select: {
@@ -74,21 +79,19 @@ const useStyles = tss.create(({ theme }) => ({
   },
 }));
 
-interface RhessysOutputsSectionProps {
-  scenarios: RhessysOutputScenario[];
-  variables: RhessysOutputVariable[];
-  isLoading: boolean;
-  /** True when the current watershed has Gate Creek-style parquet data. */
-  hasChoroplethData?: boolean;
-}
+const years = Array.from(
+  { length: GATE_CREEK_YEAR_RANGE.max - GATE_CREEK_YEAR_RANGE.min + 1 },
+  (_, i) => GATE_CREEK_YEAR_RANGE.min + i,
+);
 
-export default function RhessysOutputsSection({
-  scenarios,
-  variables,
-  isLoading,
-  hasChoroplethData = false,
-}: RhessysOutputsSectionProps) {
+export default function RhessysOutputsSection() {
   const { classes } = useStyles();
+
+  const runId = useRunId();
+
+  const { scenarios, variables, isLoading, hasChoroplethData } =
+    useRhessysOutputs(runId);
+
   const {
     layerDesired,
     enableLayerWithParams,
@@ -97,8 +100,8 @@ export default function RhessysOutputsSection({
   } = useWatershed();
 
   const params = getLayerParams(layerDesired, "rhessysOutputs");
-  const selectedScenario = params.scenario ?? "";
-  const selectedVariable = params.variable ?? "";
+  const selectedScenario = params.scenario ?? null;
+  const selectedVariable = params.variable ?? null;
   const selectedMode =
     params.mode ?? (scenarios.length > 0 ? "raster" : "choropleth");
   const selectedSpatialScale = params.spatialScale ?? "hillslope";
@@ -119,12 +122,10 @@ export default function RhessysOutputsSection({
     [selectedSpatialScale],
   );
 
-  // Shared param-update helper
-
   const currentParams = useMemo(
     () => ({
-      scenario: selectedScenario || null,
-      variable: selectedVariable || null,
+      scenario: selectedScenario,
+      variable: selectedVariable,
       mode: selectedMode,
       spatialScale: selectedSpatialScale,
       year: selectedYear,
@@ -152,8 +153,6 @@ export default function RhessysOutputsSection({
     dispatchLayerAction({ type: "TOGGLE", id: "rhessysOutputs", on: false });
   }, [dispatchLayerAction]);
 
-  // ── Pre-computed map handlers ──
-
   const handleScenarioChange = useCallback(
     (event: SelectChangeEvent) => {
       const value = event.target.value;
@@ -175,8 +174,6 @@ export default function RhessysOutputsSection({
     },
     [updateParams],
   );
-
-  // ── Dynamic choropleth handlers ──
 
   const handleSpatialScaleChange = useCallback(
     (
@@ -258,33 +255,152 @@ export default function RhessysOutputsSection({
   // Pre-computed raster maps (Victoria + Mill Creek)
   if (hasRasterData && selectedMode !== "choropleth") {
     return (
-      <RasterModeSection
-        classes={classes}
-        scenarios={scenarios}
-        availableVariables={availableVariables}
-        selectedScenario={selectedScenario}
-        selectedVariable={selectedVariable}
-        layerEnabled={layerEnabled}
-        onScenarioChange={handleScenarioChange}
-        onVariableChange={handleVariableChange}
-      />
+      <>
+        <FormControl fullWidth size="small" className={classes.formControl}>
+          <InputLabel
+            id="rhessys-outputs-scenario-label"
+            className={classes.label}
+          >
+            Scenario
+          </InputLabel>
+          <Select
+            labelId="rhessys-outputs-scenario-label"
+            id="rhessys-outputs-scenario-select"
+            value={layerEnabled && selectedScenario ? selectedScenario : "none"}
+            label="Scenario"
+            onChange={handleScenarioChange}
+            className={classes.select}
+            MenuProps={{ PaperProps: { className: classes.selectPaper } }}
+          >
+            <MenuItem value="none">None</MenuItem>
+            {scenarios.map((s) => (
+              <MenuItem key={s.id} value={s.id}>
+                {s.label}
+              </MenuItem>
+            ))}
+          </Select>
+        </FormControl>
+
+        {layerEnabled && selectedScenario && (
+          <FormControl fullWidth size="small" className={classes.formControl}>
+            <InputLabel
+              id="rhessys-outputs-variable-label"
+              className={classes.label}
+            >
+              Variable
+            </InputLabel>
+            <Select
+              labelId="rhessys-outputs-variable-label"
+              id="rhessys-outputs-variable-select"
+              value={selectedVariable || ""}
+              label="Variable"
+              onChange={handleVariableChange}
+              className={classes.select}
+              MenuProps={{ PaperProps: { className: classes.selectPaper } }}
+            >
+              {availableVariables.map((v) => (
+                <MenuItem key={v.id} value={v.id}>
+                  {v.label} ({v.units})
+                </MenuItem>
+              ))}
+            </Select>
+          </FormControl>
+        )}
+      </>
     );
   }
 
   // Dynamic choropleth (Gate Creek)
   return (
-    <ChoroplethModeSection
-      classes={classes}
-      selectedSpatialScale={selectedSpatialScale}
-      selectedScenario={selectedScenario}
-      selectedVariable={selectedVariable}
-      selectedYear={selectedYear}
-      choroplethVariables={choroplethVariables}
-      layerEnabled={layerEnabled}
-      onSpatialScaleChange={handleSpatialScaleChange}
-      onScenarioChange={handleChoroplethScenarioChange}
-      onVariableChange={handleChoroplethVariableChange}
-      onYearChange={handleYearChange}
-    />
+    <>
+      <ToggleButtonGroup
+        value={selectedSpatialScale}
+        exclusive
+        onChange={handleSpatialScaleChange}
+        size="small"
+        fullWidth
+        className={classes.toggleGroup}
+      >
+        <ToggleButton value="hillslope">Hillslope</ToggleButton>
+        <ToggleButton value="patch">Patch</ToggleButton>
+      </ToggleButtonGroup>
+
+      <FormControl fullWidth size="small" className={classes.formControl}>
+        <InputLabel
+          id="rhessys-choropleth-scenario-label"
+          className={classes.label}
+        >
+          Scenario
+        </InputLabel>
+        <Select
+          labelId="rhessys-choropleth-scenario-label"
+          id="rhessys-choropleth-scenario-select"
+          value={layerEnabled && selectedScenario ? selectedScenario : "none"}
+          label="Scenario"
+          onChange={handleChoroplethScenarioChange}
+          className={classes.select}
+          MenuProps={{ PaperProps: { className: classes.selectPaper } }}
+        >
+          <MenuItem value="none">None</MenuItem>
+          {GATE_CREEK_SCENARIOS.map((s) => (
+            <MenuItem key={s.id} value={s.id}>
+              {s.label}
+            </MenuItem>
+          ))}
+        </Select>
+      </FormControl>
+
+      {layerEnabled && selectedScenario && (
+        <>
+          <FormControl fullWidth size="small" className={classes.formControl}>
+            <InputLabel
+              id="rhessys-choropleth-variable-label"
+              className={classes.label}
+            >
+              Variable
+            </InputLabel>
+            <Select
+              labelId="rhessys-choropleth-variable-label"
+              id="rhessys-choropleth-variable-select"
+              value={selectedVariable || ""}
+              label="Variable"
+              onChange={handleChoroplethVariableChange}
+              className={classes.select}
+              MenuProps={{ PaperProps: { className: classes.selectPaper } }}
+            >
+              {choroplethVariables.map((v) => (
+                <MenuItem key={v.id} value={v.id}>
+                  {v.label} ({v.units})
+                </MenuItem>
+              ))}
+            </Select>
+          </FormControl>
+
+          <FormControl fullWidth size="small" className={classes.formControl}>
+            <InputLabel
+              id="rhessys-choropleth-year-label"
+              className={classes.label}
+            >
+              Year
+            </InputLabel>
+            <Select
+              labelId="rhessys-choropleth-year-label"
+              id="rhessys-choropleth-year-select"
+              value={String(selectedYear)}
+              label="Year"
+              onChange={handleYearChange}
+              className={classes.select}
+              MenuProps={{ PaperProps: { className: classes.selectPaper } }}
+            >
+              {years.map((y) => (
+                <MenuItem key={y} value={String(y)}>
+                  {y}
+                </MenuItem>
+              ))}
+            </Select>
+          </FormControl>
+        </>
+      )}
+    </>
   );
 }
